requirements.txt
feedparser==6.0.11
python-dotenv==1.1.0
schedule==1.2.2
